modification de supprimer ville pour ne plus supprimer les trajets dont une etape est la ville qu'on supprime

+++ CavardageModele/src/main/java/ejbs/Automate.java
@@ -73,7 +73,7 @@ public class Automate {
             for (int n : listePrix) {
                 prixMoyen += n;
             }
-            prixMoyen /= listePrix.size();
+            prixMoyen /= (float)listePrix.size();
 
             return prixMoyen;
         }
